Created config file tests in BUILD
